update on 5/18

+++ src/Employee/ControllerServlet.java
@@ -15,7 +15,10 @@ public class ControllerServlet extends HttpServlet {
     private EmployeeDAO employeeDAO;
     
     public void init() {
-    	employeeDAO = new EmployeeDAO();
+    	String name = getServletContext().getInitParameter("jdbcName");
+		String pwd = getServletContext().getInitParameter("jdbcPwd");
+		String url = getServletContext().getInitParameter("jdbcUrl");
+    	employeeDAO = new EmployeeDAO(name, pwd, url);
     }
     
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
@@ -28,27 +31,20 @@ public class ControllerServlet extends HttpServlet {
         String action = request.getServletPath();
  
         try {
-            switch (action) {
-            case "/new":
-                showNewForm(request, response);
-                break;
-            case "/insert":
-                insertEmployee(request, response);
-                break;
-            case "/delete":
-                deleteEmployee(request, response);
-                break;
-            case "/edit":
-                showEditForm(request, response);
-                break;
-            case "/update":
-                updateEmployee(request, response);
-                break;
-            default:
-                listEmployee(request, response);
-                break;
+            if(action.equals("/Employee/new")) {
+            	showNewForm(request, response);
+            }else if(action.equals("/Employee/insert")) {
+            	insertEmployee(request, response);
+            }else if(action.equals("/Employee/delete")) {
+            	deleteEmployee(request, response);
+            }else if(action.equals("/Employee/edit")) {
+            	showEditForm(request, response);
+            }else if(action.equals("/Employee/update")) {
+            	updateEmployee(request, response);
+            }else {
+            	listEmployee(request, response);
             }
-        } catch (SQLException ex) {
+        }catch (SQLException ex) {
             throw new ServletException(ex);
         }
     }
@@ -56,7 +52,7 @@ public class ControllerServlet extends HttpServlet {
     private void listEmployee(HttpServletRequest request, HttpServletResponse response)
             throws SQLException, IOException, ServletException {
         List<Employee> listEmployee = employeeDAO.listAllEmployees();
-        request.setAttribute("listEmployee", listEmployee);
+        request.setAttribute("listEmployees", listEmployee);
         RequestDispatcher dispatcher = request.getRequestDispatcher("EmployeeList.jsp");
         dispatcher.forward(request, response);
     }
@@ -72,18 +68,19 @@ public class ControllerServlet extends HttpServlet {
         int id = Integer.parseInt(request.getParameter("id"));
         Employee existingEmployee = employeeDAO.getEmployee(id);
         RequestDispatcher dispatcher = request.getRequestDispatcher("EmployeeForm.jsp");
-        request.setAttribute("book", existingEmployee);
+        request.setAttribute("employee", existingEmployee);
         dispatcher.forward(request, response);
     }
     
     private void insertEmployee(HttpServletRequest request, HttpServletResponse response)
             throws SQLException, IOException {
+    	int employeeid = Integer.parseInt(request.getParameter("employeeid"));
         String name = request.getParameter("name");
         String designation = request.getParameter("designation");
         int salary = Integer.parseInt(request.getParameter("salary"));
         String email = request.getParameter("email");
  
-        Employee newEmployee = new Employee(name, designation, salary, email);
+        Employee newEmployee = new Employee(employeeid, name, designation, salary, email);
         employeeDAO.insertEmployee(newEmployee);
         response.sendRedirect("list");
     }
@@ -91,12 +88,13 @@ public class ControllerServlet extends HttpServlet {
     private void updateEmployee(HttpServletRequest request, HttpServletResponse response)
             throws SQLException, IOException {
         int id = Integer.parseInt(request.getParameter("id"));
+        int employeeid = Integer.parseInt(request.getParameter("employeeid"));
         String name = request.getParameter("name");
         String designation = request.getParameter("designation");
         int salary = Integer.parseInt(request.getParameter("salary"));
         String email = request.getParameter("email");
  
-        Employee employee = new Employee(id, name, designation, salary, email);
+        Employee employee = new Employee(id, employeeid, name, designation, salary, email);
         employeeDAO.updateEmployee(employee);
         response.sendRedirect("list");
     }
